Refactor MoChA
